Fix bug where phantom windows would appear
This was because if a window didn't exist in the previous desktop's
windowlist, because say... it was MINIMIZED, it would add it anyway.
(After bitching about it in a print of course).

When said window got unminimized, there would now be two copies of
the same window in the list, causing the mysterious phantom window.

This fixes it, by making the function just leave if it can't remove
a window from the previous desktop list.